iox-#2052 Isolate helper for compiler compatibility

This commit extracts C++14 template structures from
`FixedPositionContainer` to a new header
`detail/fixed_position_container_helper.hpp`, resolving compile-time
differences between g++-8.4 and g++-11.4, and ensuring adherence to the
C++ standard.

In g++-8.4, nested template deduction and usage were problematic,
causing original implementation failures. Conversely, g++-11.4 processed
the same structures correctly, though reasons are not fully clear.

After this refactor, unit tests for `FixedPositionContainer` remain
successful, and the code compiles error-free on both g++-8.4 and
g++-11.4.
